chore: move ListResources method to permission manager
The ListResources method is used solely by the rebac-admin UI to list the resources that are known to JIMM. This method would need to go on a manager/business object eventually, and the permission manager made the most sense with the above in mind.
